Review feedback part 2/2: extract 1v1 match detail into subcomponent
Extract MatchDetailOneVsOne.vue from MatchDetail.vue per review feedback
to separate mode-specific rendering from the shared data layer.

The subcomponent renders the 1v1 header (player names, flags, race icons,
league/rank with glow, MMR with change), hero stats, and map info. It
receives pre-computed data as props and emits open-chat-log to the parent.
Duplicated player1/player2 computed properties are replaced by generic
player(teamIndex), mmrChange(teamIndex), and leagueName(teamIndex)
functions.

The parent remains responsible for match loading, score resolution (the
team-to-score matching system), mode routing (1v1 vs team vs FFA),
head-to-head history, and the chat log dialog.

All ovo-* styles move with the subcomponent.
