+++ src/views/MatchDetail.vue
@@ -11,249 +11,19 @@
     <v-row v-else>
       <v-col cols="12">
         <v-card tile class="pb-5 pt-1">
-          <div v-if="isOneVsOne && !loading" class="ovo-header">
-            <div class="ovo-season-bar">
-              <v-card-subtitle class="pa-0 text-uppercase opacity-100">
-                <span v-if="isGatewayNeeded">{{ $t(`gatewayNames.${gateWay}`) }} ·
-                </span>
-                <span>{{ $t(`views_matchdetail.season`) }}: {{ season }}</span>
-              </v-card-subtitle>
-              <host-icon
-                v-if="match.serverInfo && match.serverInfo.provider"
-                :host="match.serverInfo"
-                style="padding-right: 0px"
-              />
-              <div class="subicon">
-                <download-replay-icon :gameId="matchId" />
-                <v-tooltip
-                  v-if="showChatLogShortcut"
-                  location="left"
-                  content-class="w3-tooltip elevation-1"
-                >
-                  <template v-slot:activator="{ props: activatorProps }">
-                    <span v-bind="activatorProps">
-                      <v-btn
-                        class="ma-2 w3-gray-gold-text"
-                        icon
-                        variant="outlined"
-                        size="small"
-                        @click="openChatLogDialog"
-                      >
-                        <v-icon size="x-large">{{
-                          mdiChatProcessingOutline
-                        }}</v-icon>
-                      </v-btn>
-                    </span>
-                  </template>
-                  <span>View Chat Log</span>
-                </v-tooltip>
-              </div>
-            </div>
-            <div class="ovo-player-names">
-              <div class="ovo-name ovo-name--left">
-                <country-flag-extended
-                  v-if="
-                    match.teams[0].players[0].countryCode ||
-                      match.teams[0].players[0].location
-                  "
-                  :countryCode="match.teams[0].players[0].countryCode"
-                  :location="match.teams[0].players[0].location"
-                  class="mr-2"
-                />
-                <a
-                  class="text-primary cursor-pointer"
-                  @click="goToPlayer(match.teams[0].players[0].battleTag)"
-                >
-                  {{ match.teams[0].players[0].name }}
-                </a>
-              </div>
-              <player-icon
-                :race="match.teams[0].players[0].race"
-                :rndRace="match.teams[0].players[0].rndRace"
-                :big="true"
-              />
-              <div class="ovo-vs">VS</div>
-              <player-icon
-                :race="match.teams[1].players[0].race"
-                :rndRace="match.teams[1].players[0].rndRace"
-                :big="true"
-              />
-              <div class="ovo-name ovo-name--right">
-                <a
-                  class="text-primary cursor-pointer"
-                  @click="goToPlayer(match.teams[1].players[0].battleTag)"
-                >
-                  {{ match.teams[1].players[0].name }}
-                </a>
-                <country-flag-extended
-                  v-if="
-                    match.teams[1].players[0].countryCode ||
-                      match.teams[1].players[0].location
-                  "
-                  :countryCode="match.teams[1].players[0].countryCode"
-                  :location="match.teams[1].players[0].location"
-                  class="ml-2"
-                />
-              </div>
-            </div>
-            <div class="ovo-league-row">
-              <div class="ovo-league ovo-league--left">
-                <template
-                  v-if="match.teams[0].players[0].ranking?.leagueOrder != null"
-                >
-                  <router-link
-                    :to="
-                      rankingsUrl(
-                        match.teams[0].players[0],
-                        match.teams[0].players[0].ranking.leagueOrder,
-                      )
-                    "
-                    class="ovo-league-link"
-                  >
-                    <img
-                      :src="`/assets/leagueIcons/${match.teams[0].players[0].ranking.leagueOrder}.png`"
-                      class="ovo-league-icon"
-                      :class="`ovo-league-glow--${match.teams[0].players[0].ranking.leagueOrder}`"
-                    />
-                    <span class="ovo-league-name">{{ leagueNamePlayer1 }}</span>
-                    <span
-                      v-if="match.teams[0].players[0].ranking.division"
-                      class="ovo-mmr-label"
-                    >
-                      {{ match.teams[0].players[0].ranking.division }}
-                    </span>
-                    <span
-                      v-if="match.teams[0].players[0].ranking.rank"
-                      class="ovo-rank-num"
-                    >
-                      #{{ match.teams[0].players[0].ranking.rank }}
-                    </span>
-                  </router-link>
-                  <span
-                    v-if="match.teams[0].players[0].won"
-                    class="ovo-crown"
-                  >👑</span>
-                </template>
-                <template v-else>{{ $t("views_rankings.unranked") }}</template>
-              </div>
-              <div></div>
-              <div></div>
-              <div></div>
-              <div class="ovo-league ovo-league--right">
-                <template
-                  v-if="match.teams[1].players[0].ranking?.leagueOrder != null"
-                >
-                  <router-link
-                    :to="
-                      rankingsUrl(
-                        match.teams[1].players[0],
-                        match.teams[1].players[0].ranking.leagueOrder,
-                      )
-                    "
-                    class="ovo-league-link"
-                  >
-                    <img
-                      :src="`/assets/leagueIcons/${match.teams[1].players[0].ranking.leagueOrder}.png`"
-                      class="ovo-league-icon"
-                      :class="`ovo-league-glow--${match.teams[1].players[0].ranking.leagueOrder}`"
-                    />
-                    <span class="ovo-league-name">{{ leagueNamePlayer2 }}</span>
-                    <span
-                      v-if="match.teams[1].players[0].ranking.division"
-                      class="ovo-mmr-label"
-                    >
-                      {{ match.teams[1].players[0].ranking.division }}
-                    </span>
-                    <span
-                      v-if="match.teams[1].players[0].ranking.rank"
-                      class="ovo-rank-num"
-                    >
-                      #{{ match.teams[1].players[0].ranking.rank }}
-                    </span>
-                  </router-link>
-                  <span
-                    v-if="match.teams[1].players[0].won"
-                    class="ovo-crown"
-                  >👑</span>
-                </template>
-                <template v-else>{{ $t("views_rankings.unranked") }}</template>
-              </div>
-            </div>
-            <div class="ovo-mmr-row">
-              <div class="ovo-mmr ovo-mmr--left">
-                <template v-if="match.teams[0].players[0].oldMmr">
-                  {{ Math.floor(match.teams[0].players[0].oldMmr) }}
-                  <span class="ovo-mmr-label">MMR</span>
-                  <span
-                    v-if="mmrChangePlayer1 !== 0"
-                    class="number-text"
-                    :class="
-                      match.teams[0].players[0].won ? 'w3-won' : 'w3-lost'
-                    "
-                  >
-                    ({{ mmrChangePlayer1 > 0 ? "+" : ""
-                    }}{{ mmrChangePlayer1 }})
-                  </span>
-                </template>
-                <template v-else>
-                  {{
-                    $t("components_matches_playermatchinfo.calibrating")
-                  }}
-                </template>
-              </div>
-              <div></div>
-              <div></div>
-              <div></div>
-              <div class="ovo-mmr ovo-mmr--right">
-                <template v-if="match.teams[1].players[0].oldMmr">
-                  {{ Math.floor(match.teams[1].players[0].oldMmr) }}
-                  <span class="ovo-mmr-label">MMR</span>
-                  <span
-                    v-if="mmrChangePlayer2 !== 0"
-                    class="number-text"
-                    :class="
-                      match.teams[1].players[0].won ? 'w3-won' : 'w3-lost'
-                    "
-                  >
-                    ({{ mmrChangePlayer2 > 0 ? "+" : ""
-                    }}{{ mmrChangePlayer2 }})
-                  </span>
-                </template>
-                <template v-else>
-                  {{
-                    $t("components_matches_playermatchinfo.calibrating")
-                  }}
-                </template>
-              </div>
-            </div>
-          </div>
-          <div v-if="isCompleteGame && isOneVsOne">
-            <match-detail-hero-row
-              :heroes-of-winner="scoresOfWinners[0]?.heroes"
-              :heroes-of-loser="scoresOfLosers[0]?.heroes"
-              :scores-of-winner="scoresOfWinners[0]?.heroScore"
-              :scores-of-loser="scoresOfLosers[0]?.heroScore"
-              :unit-score-winner="scoresOfWinners[0]?.unitScore"
-              :unit-score-loser="scoresOfLosers[0]?.unitScore"
-              :resource-score-winner="scoresOfWinners[0]?.resourceScore"
-              :resource-score-loser="scoresOfLosers[0]?.resourceScore"
-              :duration-minutes="match.durationInSeconds / 60"
-            />
-            <div class="ovo-map-stats">
-              <div>
-                <span class="ovo-mmr-label">Map:</span>
-                {{ mapNameFromMatch(match) }}
-              </div>
-              <div>
-                <span class="ovo-mmr-label">Duration:</span>
-                {{ gameDurationLong }}
-              </div>
-              <div>
-                <span class="ovo-mmr-label">Start time:</span>
-                {{ gameStartTime }}
-              </div>
-            </div>
-          </div>
+          <match-detail-one-vs-one
+            v-if="isOneVsOne && !loading"
+            :match="match"
+            :match-id="matchId"
+            :season="season"
+            :gate-way="gateWay"
+            :is-gateway-needed="isGatewayNeeded"
+            :is-complete-game="!!isCompleteGame"
+            :scores-of-winners="scoresOfWinners"
+            :scores-of-losers="scoresOfLosers"
+            :show-chat-log-shortcut="showChatLogShortcut"
+            @open-chat-log="openChatLogDialog"
+          />
           <div v-else-if="!loading && match.teams?.length">
             <v-card-title class="justify-center">
               <v-row justify="space-around">
@@ -473,7 +243,6 @@
 
 <script lang="ts">
 import { computed, defineComponent, ref, watch, toRef } from "vue";
-import { useRouter } from "vue-router";
 import TeamMatchInfo from "@/components/matches/TeamMatchInfo.vue";
 import PlayerPerformanceOnMatch from "@/components/match-details/PlayerPerformanceOnMatch.vue";
 import MatchDetailHeroRow from "@/components/match-details/MatchDetailHeroRow.vue";
@@ -487,7 +256,6 @@ import {
   formatSecondsToDuration,
   formatTimestampStringToDateTime,
 } from "@/helpers/date-functions";
-import { formatDuration, intervalToDuration } from "date-fns";
 import { useMatchStore } from "@/store/match/store";
 import { isGatewayNeededForSeason } from "@/constants";
 import { battleTagToName } from "@/helpers/profile";
@@ -496,10 +264,7 @@ import { useOauthStore } from "@/store/oauth/store";
 import { EPermission } from "@/store/admin/permission/types";
 import AdminReplayChatLogMessages from "@/components/admin/replays/AdminReplayChatLogMessages.vue";
 import { mdiChatProcessingOutline } from "@mdi/js";
-import CountryFlagExtended from "@/components/common/CountryFlagExtended.vue";
-import PlayerIcon from "@/components/matches/PlayerIcon.vue";
-import { getProfileUrl } from "@/helpers/url-functions";
-import { leagueNameFromOrder } from "@/helpers/leagues";
+import MatchDetailOneVsOne from "@/components/match-details/MatchDetailOneVsOne.vue";
 
 export default defineComponent({
   name: "MatchDetailView",
@@ -511,8 +276,7 @@ export default defineComponent({
     DownloadReplayIcon,
     MatchHeadToHead,
     AdminReplayChatLogMessages,
-    CountryFlagExtended,
-    PlayerIcon,
+    MatchDetailOneVsOne,
   },
   props: {
     matchId: {
@@ -563,43 +327,6 @@ export default defineComponent({
       if (!match.value?.number) return false;
       return match.value.number !== 0 && match.value?.number % 1000000 === 0;
     });
-
-    const router = useRouter();
-
-    const mmrChangePlayer1 = computed<number>(() => {
-      const p = match.value?.teams?.[0]?.players?.[0];
-      if (p?.oldMmr && p?.currentMmr)
-      {return Math.floor(p.currentMmr - p.oldMmr);}
-      return 0;
-    });
-
-    const mmrChangePlayer2 = computed<number>(() => {
-      const p = match.value?.teams?.[1]?.players?.[0];
-      if (p?.oldMmr && p?.currentMmr)
-      {return Math.floor(p.currentMmr - p.oldMmr);}
-      return 0;
-    });
-
-    const gameDurationLong = computed<string>(() => formatDuration(intervalToDuration({ start: 0, end: match.value.durationInSeconds * 1000 })));
-    const gameStartTime = computed<string>(() => new Date(match.value.startTime).toLocaleString());
-
-    const leagueNamePlayer1 = computed(() => leagueNameFromOrder(match.value?.teams?.[0]?.players?.[0]?.ranking?.leagueOrder));
-    const leagueNamePlayer2 = computed(() => leagueNameFromOrder(match.value?.teams?.[1]?.players?.[0]?.ranking?.leagueOrder));
-
-    function rankingsUrl(player: PlayerInTeam, league?: number) {
-      const query: Record<string, string> = {
-        season: String(season.value),
-        gateway: String(match.value.gateWay),
-        gamemode: String(match.value.gameMode),
-        playerId: player.battleTag,
-      };
-      if (league != null) query.league = String(league);
-      return { path: "/rankings", query };
-    }
-
-    function goToPlayer(battleTag: string) {
-      router.push({ path: getProfileUrl(battleTag) });
-    }
 
     const playerScores = computed<PlayerScore[]>(() => {
       const { playerScores, match } = matchStore.matchDetail;
@@ -903,14 +630,6 @@ export default defineComponent({
       showChatLogShortcut,
       openChatLogDialog,
       mdiChatProcessingOutline,
-      mmrChangePlayer1,
-      mmrChangePlayer2,
-      leagueNamePlayer1,
-      leagueNamePlayer2,
-      rankingsUrl,
-      gameDurationLong,
-      gameStartTime,
-      goToPlayer,
     };
   },
 });
@@ -946,183 +665,5 @@ export default defineComponent({
   position: absolute;
   top: 8px;
   right: 8px;
-}
-
-.ovo-header {
-  padding: 8px 16px;
-  position: relative;
-}
-
-.ovo-season-bar {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  margin-bottom: 8px;
-
-  .subicon {
-    position: static;
-    margin-left: auto;
-    flex-direction: row;
-  }
-}
-
-.ovo-player-names {
-  display: grid;
-  grid-template-columns: 1fr auto auto auto 1fr;
-  align-items: center;
-  grid-column-gap: 10px;
-  margin-bottom: 4px;
-}
-
-.ovo-name {
-  font-weight: bold;
-  font-size: 1.5em;
-
-  @media (max-width: 850px) {
-    font-size: 1.1em;
-  }
-  overflow: hidden;
-
-  a {
-    text-decoration: none;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    display: block;
-    &:hover {
-      text-decoration: underline;
-    }
-  }
-}
-
-.ovo-name--left {
-  text-align: right;
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-}
-
-.ovo-name--right {
-  text-align: left;
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-}
-
-.ovo-vs {
-  font-size: 1.2em;
-  font-weight: bold;
-  padding: 0 8px;
-}
-
-.ovo-league-row {
-  display: grid;
-  grid-template-columns: 1fr auto auto auto 1fr;
-  align-items: center;
-  grid-column-gap: 10px;
-  margin-top: 16px;
-}
-
-.ovo-league {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-}
-
-.ovo-league--left {
-  justify-content: flex-end;
-}
-
-.ovo-league--right {
-  justify-content: flex-start;
-  padding-left: 12px;
-}
-
-.ovo-league-icon {
-  width: 32px;
-  height: 32px;
-
-  &.ovo-league-glow--0 { filter: drop-shadow(0 0 5px rgba(255, 215, 0, 0.5)); }
-  &.ovo-league-glow--1 { filter: drop-shadow(0 0 5px rgba(180, 130, 255, 0.5)); }
-  &.ovo-league-glow--2 { filter: drop-shadow(0 0 5px rgba(0, 200, 200, 0.5)); }
-  &.ovo-league-glow--3 { filter: drop-shadow(0 0 5px rgba(100, 180, 255, 0.5)); }
-  &.ovo-league-glow--4 { filter: drop-shadow(0 0 5px rgba(180, 220, 255, 0.5)); }
-  &.ovo-league-glow--5 { filter: drop-shadow(0 0 5px rgba(255, 215, 0, 0.5)); }
-  &.ovo-league-glow--6 { filter: drop-shadow(0 0 5px rgba(192, 192, 192, 0.5)); }
-  &.ovo-league-glow--7 { filter: drop-shadow(0 0 5px rgba(205, 127, 50, 0.5)); }
-  &.ovo-league-glow--8 { filter: drop-shadow(0 0 5px rgba(100, 200, 100, 0.5)); }
-}
-
-.ovo-league-name {
-  @media (max-width: 850px) {
-    display: none;
-  }
-}
-
-@media (max-width: 850px) {
-  .ovo-league-icon {
-    width: 40px;
-    height: 40px;
-  }
-}
-
-.ovo-rank-num {
-  opacity: 1;
-}
-
-.ovo-league-link {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  text-decoration: none;
-  color: inherit;
-  cursor: pointer;
-  transition: box-shadow 0.15s;
-
-  &:hover {
-    box-shadow: 0 0 10px rgba(var(--v-theme-on-surface), 0.3);
-  }
-}
-
-.ovo-mmr-row {
-  display: grid;
-  grid-template-columns: 1fr 10px 141px 10px 1fr;
-  align-items: center;
-  margin-top: 15px;
-  margin-bottom: 1px;
-}
-
-.ovo-mmr--left {
-  text-align: right;
-  font-size: 0.85em;
-  opacity: 0.8;
-}
-
-.ovo-mmr--right {
-  text-align: left;
-  font-size: 0.85em;
-  opacity: 0.8;
-}
-
-.ovo-mmr-label {
-  opacity: 0.6;
-}
-
-.ovo-winner-icon {
-  width: 40px;
-  text-align: center;
-}
-
-.ovo-crown {
-  font-size: 1.5em;
-  padding-right: 13px;
-}
-
-.ovo-map-stats {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 2px;
-  padding: 16px 0 8px;
 }
 </style>
